+++ data/src/main/java/daily/dayo/data/datasource/remote/post/PostApiService.kt
@@ -25,7 +25,7 @@ interface PostApiService {
 
     @POST("/api/v1/posts/{postId}/edit")
     suspend fun requestEditPost(
-        @Path("postId") postId: Int,
+        @Path("postId") postId: Long,
         @Body body: EditPostRequest
     ): NetworkResponse<EditPostResponse>
 
@@ -42,10 +42,10 @@ interface PostApiService {
     suspend fun requestDayoPickPostListCategory(@Path("category") category: Category): NetworkResponse<DayoPickPostListResponse>
 
     @GET("/api/v1/posts/{postId}")
-    suspend fun requestPostDetail(@Path("postId") postId: Int): NetworkResponse<DetailPostResponse>
+    suspend fun requestPostDetail(@Path("postId") postId: Long): NetworkResponse<DetailPostResponse>
 
     @POST("/api/v1/posts/delete/{postId}")
-    suspend fun requestDeletePost(@Path("postId") postId: Int): NetworkResponse<Void>
+    suspend fun requestDeletePost(@Path("postId") postId: Long): NetworkResponse<Void>
 
     @GET("/api/v1/posts/feed/list")
     suspend fun requestAllFeedList(@Query("end") end: Int): NetworkResponse<ListFeedResponse>
