+++ domain/src/main/java/daily/dayo/domain/model/Bookmark.kt
@@ -1,10 +1,10 @@
 package daily.dayo.domain.model
 
 data class BookmarkPost(
-    val postId: Int,
+    val postId: Long,
     val thumbnailImage: String
 )
 data class BookmarkPostResponse(
     val memberId: String,
-    val postId: Int
+    val postId: Long
 )
