+++ domain/src/main/java/daily/dayo/domain/repository/CommentRepository.kt
@@ -6,8 +6,8 @@ import daily.dayo.domain.model.NetworkResponse
 
 interface CommentRepository {
 
-    suspend fun requestCreatePostComment(contents: String, postId: Int, mentionList: List<MentionUser>): NetworkResponse<Int>
-    suspend fun requestCreatePostCommentReply(commentId: Long, contents: String, postId: Int, mentionList: List<MentionUser>): NetworkResponse<Int>
-    suspend fun requestPostComment(postId: Int): NetworkResponse<Comments>
+    suspend fun requestCreatePostComment(contents: String, postId: Long, mentionList: List<MentionUser>): NetworkResponse<Int>
+    suspend fun requestCreatePostCommentReply(commentId: Long, contents: String, postId: Long, mentionList: List<MentionUser>): NetworkResponse<Int>
+    suspend fun requestPostComment(postId: Long): NetworkResponse<Comments>
     suspend fun requestDeletePostComment(commentId: Long): NetworkResponse<Void>
 }
