+++ domain/src/main/java/daily/dayo/domain/usecase/like/RequestPostLikeUsersUseCase.kt
@@ -6,6 +6,6 @@ import javax.inject.Inject
 class RequestPostLikeUsersUseCase @Inject constructor(
     private val heartRepository: HeartRepository
 ) {
-    suspend operator fun invoke(postId: Int) =
+    suspend operator fun invoke(postId: Long) =
         heartRepository.requestPostLikeUsers(postId = postId)
 }
